% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{proccess_video_yolo}
\alias{proccess_video_yolo}
\title{YOLO back-end for emphazis}
\usage{
proccess_video_yolo(video_path, subject_model, fps, envir = parent.frame())
}
\arguments{
\item{video_path}{Path to video file.}

\item{subject_model}{Path to model file of folder.}

\item{fps}{FPS value.}

\item{envir}{Environment for progress bar.}
}
\description{
YOLO back-end for \code{emphazis} tracking functionality.
}
